pkg_install: Fix upgrade issues with dependencies.
Previously there were a couple of cases where upgrades that themselves
pulled in upgraded dependencies weren't handled properly.

Firstly, the list of dependencies held by packages during recursive
upgrades was not refreshed after dependencies were themselves upgraded,
leading to failures attempting to read +REQUIRED_BY files in package
directories that no longer exist ("registration is incomplete!").  We
now only perform the package match after the upgrades have completed.

Secondly, when removing packages, the dependency match used by the
package in question may no longer match what is currently installed, for
example if we are upgrading a set of packages simultaneously past a
specific version match.  As it's not possible to update multiple
packages atomically there will be a mismatch between the +CONTENTS
@pkgdep of the still-installed package compared with the updated version
of the now-upgraded package.  To work around this, any specific matches
for example >=X<Y are converted to general globs (-[0-9]*) so that any
installed package version will be matched.  As there should be no
duplicate packages installed using the same PKGBASE, there should in
theory be no possibility of matching the wrong package.

While here, hide the warning about dependencies not being fulfilled
correctly behind ForceDepending, as the whole point of using that mode
is to ignore such issues with the assumption that the final state after
updating will be correct.