lang/mono: allow the SUBST block fix-prefix to not affect all files
The list of files is generated via find(1) and contains each file
individually.  Some of these files are modified, some aren't.  Those
files that aren't modified are redundant, but since they are not listed
explicitly in the package Makefile, there is no superfluous code.
Avoiding such superfluous code is the whole goal of SUBST_NOOP_OK.